requirements.txt
matplotlib
numpy
pandas
python-dotenv
scikit-learn
seaborn
sentence-transformers
slack-sdk
streamlit
torch
